backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
web3==6.11.3
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9